+++ crypto/src/commonMain/kotlin/at/asitplus/awesn1/crypto/Pkcs1RsaPrivateKeyInfo.kt
@@ -3,11 +3,23 @@
 
 package at.asitplus.awesn1.crypto
 
+import at.asitplus.awesn1.Asn1Element
 import at.asitplus.awesn1.Asn1Integer
+import at.asitplus.awesn1.Asn1Null
+import at.asitplus.awesn1.Asn1OctetString
+import at.asitplus.awesn1.ObjectIdentifier
 import at.asitplus.awesn1.serialization.Asn1Tag
 import at.asitplus.awesn1.WithPemLabel
 import at.asitplus.awesn1.PemLabelSpec
+import at.asitplus.awesn1.encoding.Asn1
+import at.asitplus.awesn1.runRethrowing
+import at.asitplus.awesn1.serialization.DER
+import at.asitplus.awesn1.serialization.Der
+import at.asitplus.awesn1.serialization.decodeFromDer
+import at.asitplus.awesn1.serialization.decodeFromTlv
+import at.asitplus.awesn1.serialization.encodeToTlv
 import kotlinx.serialization.Serializable
+import kotlinx.serialization.encodeToByteArray
 
 /**
  *
@@ -61,6 +73,27 @@ data class Pkcs1RsaPrivateKeyInfo(
     companion object : PemLabelSpec<Pkcs1RsaPrivateKeyInfo> {
         const val PEM_LABEL = "RSA PRIVATE KEY"
         override val canonicalPemLabel: String get() = PEM_LABEL
+        private val RSA_ENCRYPTION_OID = ObjectIdentifier("1.2.840.113549.1.1.1")
+
+        fun of(privateKeyInfo: Pkcs8PrivateKeyInfo, der: Der = DER) = runRethrowing {
+            require(privateKeyInfo.algorithmOid == RSA_ENCRYPTION_OID)
+                { "Pkcs8PrivateKeyInfo is not an RSA private key" }
+            require(privateKeyInfo.algorithmParameters == Asn1Null)
+                { "RSA SubjectPublicKeyInfo must contain NULL params" }
+            der.decodeFromTlv<Pkcs1RsaPrivateKeyInfo>(
+                privateKeyInfo.privateKey.asEncapsulatingOctetString().element)
+        }
+
+        operator fun Pkcs8PrivateKeyInfo.Companion.invoke(
+                privateKey: Pkcs1RsaPrivateKeyInfo, attributes: Set<Asn1Element>? = null, der: Der = DER
+        ) = runRethrowing {
+            Pkcs8PrivateKeyInfo(
+                version = Pkcs8PrivateKeyInfo.Version.V1,
+                privateKeyAlgorithm = X509AlgorithmIdentifier(RSA_ENCRYPTION_OID, Asn1Null),
+                privateKey = Asn1OctetString(der.encodeToByteArray(privateKey)),
+                attributes = attributes,
+            )
+        }
     }
 }