+++ crypto/src/commonMain/kotlin/at/asitplus/awesn1/crypto/Pkcs8PrivateKeyInfo.kt
@@ -4,7 +4,8 @@
 package at.asitplus.awesn1.crypto
 
 import at.asitplus.awesn1.*
-import at.asitplus.awesn1.encoding.Asn1
+import at.asitplus.awesn1.crypto.Pkcs1RsaPrivateKeyInfo.Companion.invoke
+import at.asitplus.awesn1.crypto.Sec1EcPrivateKeyInfo.Companion.invoke
 import at.asitplus.awesn1.serialization.Asn1Tag
 import at.asitplus.awesn1.serialization.DER
 import at.asitplus.awesn1.serialization.decodeFromTlv
@@ -26,6 +27,8 @@ import kotlinx.serialization.Serializable
  * PrivateKey ::= OCTET STRING
  * Attributes ::= SET OF Attribute
  * ```
+ * @see Pkcs1RsaPrivateKeyInfo.of
+ * @see Sec1EcPrivateKeyInfo.of
  */
 @Serializable
 data class Pkcs8PrivateKeyInfo(
@@ -40,17 +43,15 @@ data class Pkcs8PrivateKeyInfo(
 
     override val pemLabel: String get() = PEM_LABEL_PRIVATE_KEY
 
-    @Throws(Asn1Exception::class)
-    fun decodeRsaPrivateKey(): Pkcs1RsaPrivateKeyInfo =
-        DER.decodeFromTlv(privateKey.asEncapsulatingOctetString().decodeRethrowing { next() })
+    @Deprecated("Moved to a more suitable location", ReplaceWith("Pkcs1RsaPrivateKeyInfo.of(this)"))
+    fun decodeRsaPrivateKey() =
+        Pkcs1RsaPrivateKeyInfo.of(this)
 
-    @Throws(Asn1Exception::class)
+    @Deprecated("Moved to a more suitable location", ReplaceWith("Sec1EcPrivateKeyInfo.of(this)"))
     fun decodeEcPrivateKey(): Sec1EcPrivateKeyInfo =
-        DER.decodeFromTlv(privateKey.asEncapsulatingOctetString().decodeRethrowing { next() })
+        Sec1EcPrivateKeyInfo.of(this)
 
     companion object : PemLabelSpec<Pkcs8PrivateKeyInfo> {
-        private val RSA_ENCRYPTION_OID = ObjectIdentifier("1.2.840.113549.1.1.1")
-        private val EC_PUBLIC_KEY_OID = ObjectIdentifier("1.2.840.10045.2.1")
 
         const val PEM_LABEL_PRIVATE_KEY = "PRIVATE KEY"
         const val PEM_LABEL_RSA_PRIVATE_KEY = Pkcs1RsaPrivateKeyInfo.PEM_LABEL
@@ -60,28 +61,15 @@ data class Pkcs8PrivateKeyInfo(
         override val alternativePemLabels: Set<String> =
             setOf(PEM_LABEL_RSA_PRIVATE_KEY, PEM_LABEL_EC_PRIVATE_KEY)
 
+        @Deprecated("Moved to an extension on Pkcs1RsaPrivateKeyInfo's companion",
+            ReplaceWith("Pkcs8PrivateKeyInfo(privateKey, attributes)"))
+        fun rsa(privateKey: Pkcs1RsaPrivateKeyInfo, attributes: Set<Asn1Element>? = null) =
+            this(privateKey, attributes)
 
-        fun rsa(privateKey: Pkcs1RsaPrivateKeyInfo, attributes: Set<Asn1Element>? = null): Pkcs8PrivateKeyInfo =
-            Pkcs8PrivateKeyInfo(
-                version = Version.V1,
-                privateKeyAlgorithm = X509AlgorithmIdentifier(RSA_ENCRYPTION_OID, Asn1.Null()),
-                privateKey = Asn1.OctetStringEncapsulating { +DER.encodeToTlv(privateKey) },
-                attributes = attributes,
-            )
-
-        fun ec(
-            sec1Key: Sec1EcPrivateKeyInfo,
-            curveOid: ObjectIdentifier?,
-            attributes: Set<Asn1Element>? = null,
-        ): Pkcs8PrivateKeyInfo = Pkcs8PrivateKeyInfo(
-            version = Version.V1,
-            privateKeyAlgorithm = X509AlgorithmIdentifier(
-                EC_PUBLIC_KEY_OID,
-                curveOid?.encodeToTlv(),
-            ),
-            privateKey = Asn1.OctetStringEncapsulating { +DER.encodeToTlv(sec1Key) },
-            attributes = attributes,
-        )
+        @Deprecated("Moved to an extension on Sec1EcPrivateKeyInfo's companion",
+            ReplaceWith("Pkcs8PrivateKeyInfo(sec1Key, curveOid, attributes)"))
+        fun ec(sec1Key: Sec1EcPrivateKeyInfo, curveOid: ObjectIdentifier?, attributes: Set<Asn1Element>? = null) =
+            this(sec1Key, curveOid, attributes)
     }
 
     /**