+++ crypto/src/commonMain/kotlin/at/asitplus/awesn1/crypto/X509AlgorithmIdentifier.kt
@@ -22,6 +22,7 @@ import kotlin.native.HiddenFromObjC
  * ```
  *
  * As a transparent [WrappedElement], an identifier can be added directly to an ASN.1 builder with unary `+`.
+ * @see RsaSsaPssParams.of
  */
 @JvmInline
 @Serializable